Implement linked list file system

+++ src/common/fileSystem.c
@@ -4,7 +4,7 @@
 allocPointer_t openFileByStringAlloc(allocPointer_t stringAlloc) {
     heapMemoryOffset_t tempAddress = getAllocDataAddress(stringAlloc);
     heapMemoryOffset_t tempSize = getAllocSize(stringAlloc);
-    return openFile(tempAddress, tempSize);
+    return openFile(tempAddress, (uint8_t)tempSize);
 }
